move storage utils file to importer
